pyproject.toml
[tool.poetry]
name = "tghtml"
version = "1.1.5"
description = "Simple tool for parse common HTML to Telegram HTML"
authors = ["Daniel Zakharov <gzdan734@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.1"
readability-lxml = "^0.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
